# swagger/docs.yml
openapi: 3.0.1
info:
  title: Library System
  version: "1.0"
servers:
  - url: http://localhost:8080
paths:
  /api/v1/register:
    post:
      summary: Register a new user
      tags:
        - Identity Provider API
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserCreateRequest"
      responses:
        "200":
          description: Successful registration
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserCreateResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/v1/authorize:
    post:
      summary: Authorize a user
      tags:
        - Identity Provider API
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuthRequest"
      responses:
        "200":
          description: Successful authorization
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/v1/libraries:
    get:
      summary: Получить список библиотек в городе
      tags:
        - Gateway API
      security:
        - bearerAuth: [ ]
      parameters:
        - name: page
          in: query
          required: false
          schema:
            type: number
            minimum: 0
        - name: size
          in: query
          required: false
          schema:
            type: number
            minimum: 1
            maximum: 100
        - name: city
          in: query
          required: true
          description: Город
          schema:
            type: string
      responses:
        "200":
          description: Список библиотек в городе
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LibraryPaginationResponse"
        "401":
          description: Ошибка авторизации
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/v1/libraries/{libraryUid}/books:
    get:
      summary: Получить список книг в выбранной библиотеке
      tags:
        - Gateway API
      security:
        - bearerAuth: [ ]
      parameters:
        - name: page
          in: query
          required: false
          schema:
            type: number
            minimum: 0
        - name: size
          in: query
          required: false
          schema:
            type: number
            minimum: 1
            maximum: 100
        - name: showAll
          in: query
          required: false
          schema:
            type: boolean
        - name: libraryUid
          in: path
          required: true
          description: UUID библиотеки
          schema:
            type: string
      responses:
        "200":
          description: Список книг библиотеке
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LibraryBookPaginationResponse"
        "401":
          description: Ошибка авторизации
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/v1/reservations:
    get:
      summary: Получить информацию по всем взятым в прокат книгам пользователя
      tags:
        - Gateway API
      security:
        - bearerAuth: [ ]
      responses:
        "200":
          description: Информация по всем взятым в прокат книгам
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BookReservationResponse"
        "401":
          description: Ошибка авторизации
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    post:
      summary: Взять книгу в библиотеке
      tags:
        - Gateway API
      security:
        - bearerAuth: [ ]
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TakeBookRequest"
      responses:
        "200":
          description: Информация о бронировании
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TakeBookResponse"
        "400":
          description: Ошибка валидации данных
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValidationErrorResponse"
        "401":
          description: Ошибка авторизации
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/v1/reservations/{reservationUid}/return:
    post:
      summary: Вернуть книгу
      tags:
        - Gateway API
      security:
        - bearerAuth: [ ]
      parameters:
        - name: reservationUid
          in: path
          description: UUID бронирования
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReturnBookRequest"
      responses:
        "204":
          description: Книга успешно возвращена
        "401":
          description: Ошибка авторизации
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Бронирование не найдено
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    AuthRequest:
      type: object
      required:
        - username
        - password
      properties:
        username:
          type: string
        password:
          type: string
    AuthResponse:
      type: object
      properties:
        expiresIn:
          type: integer
        accessToken:
          type: string
    UserCreateRequest:
      type: object
      required:
        - username
        - password
        - email
      properties:
        username:
          type: string
        password:
          type: string
        email:
          type: string
    LibraryPaginationResponse:
      type: object
      example:
        {
          "page": 1,
          "pageSize": 1,
          "totalElements": 1,
          "items": [
            {
              "libraryUid": "83575e12-7ce0-48ee-9931-51919ff3c9ee",
              "name": "Библиотека имени 7 Непьющих",
              "address": "2-я Бауманская ул., д.5, стр.1",
              "city": "Москва"
            }
          ]
        }
      properties:
        page:
          type: number
          description: Номер страницы
        pageSize:
          type: number
          description: Количество элементов на странице
        totalElements:
          type: number
          description: Общее количество элементов
        items:
          type: array
          items:
            $ref: "#/components/schemas/LibraryResponse"
    LibraryResponse:
      type: object
      example:
        {
          "libraryUid": "83575e12-7ce0-48ee-9931-51919ff3c9ee",
          "name": "Библиотека имени 7 Непьющих",
          "address": "2-я Бауманская ул., д.5, стр.1",
          "city": "Москва"
        }
      properties:
        libraryUid:
          type: string
          description: UUID библиотеки
          format: uuid
        name:
          type: string
          description: Название библиотеки
        address:
          type: string
          description: Адрес библиотеки
        city:
          type: string
          description: Город, в котором находится библиотека
    LibraryBookPaginationResponse:
      type: object
      example:
        {
          "page": 1,
          "pageSize": 1,
          "totalElements": 1,
          "items": [
            {
              "bookUid": "f7cdc58f-2caf-4b15-9727-f89dcc629b27",
              "name": "Краткий курс C++ в 7 томах",
              "author": "Бьерн Страуструп",
              "genre": "Научная фантастика",
              "condition": "EXCELLENT",
              "availableCount": 1
            }
          ]
        }
      properties:
        page:
          type: number
          description: Номер страницы
        pageSize:
          type: number
          description: Количество элементов на странице
        totalElements:
          type: number
          description: Общее количество элементов
        items:
          type: array
          items:
            $ref: "#/components/schemas/LibraryBookResponse"
    LibraryBookResponse:
      type: object
      example:
        {
          "bookUid": "f7cdc58f-2caf-4b15-9727-f89dcc629b27",
          "name": "Краткий курс C++ в 7 томах",
          "author": "Бьерн Страуструп",
          "genre": "Научная фантастика",
          "condition": "EXCELLENT",
          "availableCount": 1
        }
      properties:
        bookUid:
          type: string
          description: UUID книги
          format: uuid
        name:
          type: string
          description: Название книги
        author:
          type: string
          description: Автор
        genre:
          type: string
          description: Жанр
        condition:
          type: string
          description: Состояние книги
          enum:
            - EXCELLENT
            - GOOD
            - BAD
        availableCount:
          type: number
          description: Количество книг, доступных для аренды в библиотеке
    BookReservationResponse:
      type: object
      example:
        {
          "reservationUid": "f464ca3a-fcf7-4e3f-86f0-76c7bba96f72",
          "status": "RENTED",
          "startDate": "2021-10-09",
          "tillDate": "2021-10-11",
          "book": {
            "bookUid": "f7cdc58f-2caf-4b15-9727-f89dcc629b27",
            "name": "Краткий курс C++ в 7 томах",
            "author": "Бьерн Страуструп",
            "genre": "Научная фантастика"
          },
          "library": {
            "libraryUid": "83575e12-7ce0-48ee-9931-51919ff3c9ee",
            "name": "Библиотека имени 7 Непьющих",
            "address": "2-я Бауманская ул., д.5, стр.1",
            "city": "Москва"
          }
        }
      properties:
        reservationUid:
          type: string
          description: UUID бронирования
          format: uuid
        status:
          type: string
          description: Статус бронирования книги
          enum:
            - RENTED
            - RETURNED
            - EXPIRED
        startDate:
          type: string
          description: Дата начала бронирования
          format: ISO 8601
        tillDate:
          type: string
          description: Дата окончания бронирования
          format: ISO 8601
        book:
          $ref: "#/components/schemas/BookInfo"
        library:
          $ref: "#/components/schemas/LibraryResponse"
    TakeBookRequest:
      type: object
      example:
        {
          "bookUid": "f7cdc58f-2caf-4b15-9727-f89dcc629b27",
          "libraryUid": "83575e12-7ce0-48ee-9931-51919ff3c9ee",
          "tillDate": "2021-10-11"
        }
      properties:
        bookUid:
          type: string
          description: UUID книги
          format: uuid
        libraryUid:
          type: string
          description: UUID библиотеки
          format: uuid
        tillDate:
          type: string
          description: Дата окончания бронирования
          format: ISO 8601
    TakeBookResponse:
      type: object
      example:
        {
          "reservationUid": "f464ca3a-fcf7-4e3f-86f0-76c7bba96f72",
          "status": "RENTED",
          "startDate": "2021-10-09",
          "tillDate": "2021-10-11",
          "book": {
            "bookUid": "f7cdc58f-2caf-4b15-9727-f89dcc629b27",
            "name": "Краткий курс C++ в 7 томах",
            "author": "Бьерн Страуструп",
            "genre": "Научная фантастика",
          },
          "library": {
            "libraryUid": "83575e12-7ce0-48ee-9931-51919ff3c9ee",
            "name": "Библиотека имени 7 Непьющих",
            "address": "2-я Бауманская ул., д.5, стр.1",
            "city": "Москва"
          },
          "rating": {
            "stars": 75
          }
        }
      properties:
        reservationUid:
          type: string
          description: UUID бронирования
          format: uuid
        status:
          type: string
          description: Статус бронирования книги
          enum:
            - RENTED
            - RETURNED
            - EXPIRED
            - LOST
        startDate:
          type: string
          description: Дата начала бронирования
          format: ISO 8601
        tillDate:
          type: string
          description: Дата окончания бронирования
          format: ISO 8601
        book:
          $ref: "#/components/schemas/BookInfo"
        library:
          $ref: "#/components/schemas/LibraryResponse"
        rating:
          $ref: "#/components/schemas/UserRatingResponse"
    ReturnBookRequest:
      type: object
      example:
        {
          "condition": "EXCELLENT",
          "date": "2021-10-11"
        }
      properties:
        condition:
          type: string
          description: Состояние книги
          enum:
            - EXCELLENT
            - GOOD
            - BAD
        date:
          type: string
          description: Дата возврата
          format: ISO 8601
    UserRatingResponse:
      type: object
      example:
        {
          "stars": 75
        }
      properties:
        stars:
          type: number
          description: Количество здесь у пользователя
          minimum: 0
          maximum: 100
...
    BookInfo:
      type: object
      example:
        {
          "bookUid": "f7cdc58f-2caf-4b15-9727-f89dcc629b27",
          "name": "Краткий курс C++ в 7 томах",
          "author": "Бьерн Страуструп",
          "genre": "Научная фантастика",
        }
      properties:
        bookUid:
          type: string
          description: UUID книги
          format: uuid
        name:
          type: string
          description: Название книги
        author:
          type: string
          description: Автор
        genre:
          type: string
          description: Жанр
    ErrorDescription:
      type: object
      properties:
        field:
          type: string
        error:
          type: string
    ErrorResponse:
      type: object
      properties:
        message:
          type: string
          description: Информация об ошибке
    ValidationErrorResponse:
      type: object
      properties:
        message:
          type: string
          description: Информация об ошибке
        errors:
          type: array
          description: Массив полей с описанием ошибки
          items:
            $ref: "#/components/schemas/ErrorDescription"